+++ tests/test_config.py
@@ -1,6 +1,6 @@
 import os
 from unittest import TestCase
-from unittest.mock import PropertyMock, patch
+from unittest.mock import MagicMock, PropertyMock, patch
 
 try:
     from importlib.resources import path
@@ -8,8 +8,10 @@ try:
 except ImportError:
     from importlib_resources import path
 
+from pathlib import Path
+
 from environs import Env
-from path import Path
+from platformdirs import PlatformDirs
 from yaml.parser import ParserError
 
 from dakara_base.config import (
@@ -279,15 +281,26 @@ class SetLoglevelTestCase(TestCase):
         mocked_set_level.assert_called_with("INFO")
 
 
-@patch.object(Path, "copyfile")
+# fix required for Python 3.8 as you seemingly cannot use an invalid path as a
+# context manager
+def mock_context_manager(return_value):
+    mock = MagicMock()
+    mock.__enter__.return_value = return_value
+
+    return mock
+
+
+@patch("dakara_base.config.shutil.copyfile")
 @patch.object(Path, "exists")
-@patch.object(Path, "mkdir_p")
-@patch(
-    "dakara_base.directory.AppDirsPath.user_config_dir",
+@patch.object(Path, "mkdir")
+@patch.object(
+    PlatformDirs,
+    "user_config_dir",
     new_callable=PropertyMock(return_value=Path("path") / "to" / "directory"),
 )
 @patch(
     "dakara_base.config.path",
+    return_value=mock_context_manager(Path("path") / "to" / "source"),
 )
 class CreateConfigFileTestCase(TestCase):
     """Test the config file creator."""
@@ -296,7 +309,7 @@ class CreateConfigFileTestCase(TestCase):
         self,
         mocked_path,
         mocked_user_config_dir,
-        mocked_mkdir_p,
+        mocked_mkdir,
         mocked_exists,
         mocked_copyfile,
     ):
@@ -310,10 +323,11 @@ class CreateConfigFileTestCase(TestCase):
 
         # assert the call
         mocked_path.assert_called_with("module.resources", "config.yaml")
-        mocked_mkdir_p.assert_called_with()
+        mocked_mkdir.assert_called_with(parents=True, exists_ok=True)
         mocked_exists.assert_called_with()
         mocked_copyfile.assert_called_with(
-            Path("path") / "to" / "directory" / "config.yaml"
+            Path("path") / "to" / "source",
+            Path("path") / "to" / "directory" / "config.yaml",
         )
 
         # assert the logs
@@ -332,7 +346,7 @@ class CreateConfigFileTestCase(TestCase):
         mocked_input,
         mocked_path,
         mocked_user_config_dir,
-        mocked_mkdir_p,
+        mocked_mkdir,
         mocked_exists,
         mocked_copyfile,
     ):
@@ -358,7 +372,7 @@ class CreateConfigFileTestCase(TestCase):
         mocked_input,
         mocked_path,
         mocked_user_config_dir,
-        mocked_mkdir_p,
+        mocked_mkdir,
         mocked_exists,
         mocked_copyfile,
     ):
@@ -379,7 +393,7 @@ class CreateConfigFileTestCase(TestCase):
         mocked_input,
         mocked_path,
         mocked_user_config_dir,
-        mocked_mkdir_p,
+        mocked_mkdir,
         mocked_exists,
         mocked_copyfile,
     ):
@@ -391,5 +405,6 @@ class CreateConfigFileTestCase(TestCase):
         mocked_exists.assert_not_called()
         mocked_input.assert_not_called()
         mocked_copyfile.assert_called_with(
-            Path("path") / "to" / "directory" / "config.yaml"
+            Path("path") / "to" / "source",
+            Path("path") / "to" / "directory" / "config.yaml",
         )